requirements.txt
flask>=2.0.3
python-dotenv>=0.19.0
zhipuai>=1.0.2